+++ libs/gamepad/sim/state.ts
