--- libs/joystick/joystick.ts
@@ -1,17 +1,0 @@
-//% icon="\uf11b" color="#777777"
-namespace joystick {
-    //% shim=joystick::setButton
-    function setButton(index: int32, down: boolean): void {
-
-    }
-
-    //% shim=joystick::move
-    function move(index: int32, x: int32, y: int32): void{
-
-    }
-
-    //% shim=joystick::setThrottle
-    function setThrottle(index: int32, value: int32): void {
-
-    }
-}
